Split list and deletecollection handlers into seperate files
This PR splits the resource list and delete collection handlers into
separate files.

It does not include any change to the codebase.
